fix new project errors

+++ smartHouse/src/main/java/sensor/sensorClass/CO2Sensor.java
@@ -22,7 +22,7 @@ public class CO2Sensor implements InterfaceSensors {
         return listeOfValue;
     }
 
-    public float getLastValue(){
+    public double getLastValue(){
         //float value = db.getLastValue();
         int value=1; // suppsr
         return value;
